LayoutTests/http/tests/w3c/dom/nodes/Document-constructor-expected.txt


FAIL new Document(): interfaces assert_false: Should not be an XMLDocument expected false got true
PASS new Document(): children 
FAIL new Document(): metadata assert_equals: expected "about:blank" but got ""
FAIL new Document(): URL parsing assert_equals: expected "http://example.org/?%C3%A4" but got "http://example.org/?ä"
